refactor: move constants in seed
