+++ packages/cli/src/lib/services/seed/collections/collection.ts
@@ -1,5 +1,6 @@
 import { Inject, Service } from 'typedi';
-import { COLLECTION, LOGGER, META } from '../../../constants';
+import { LOGGER } from '../../../constants';
+import { COLLECTION, META } from '../constants';
 import pino from 'pino';
 import { getChildLogger } from '../../../helpers';
 import { SeedMeta, SeedData } from '../interfaces';
