+++ packages/cli/src/lib/services/seed/collections/data-client.ts
@@ -4,7 +4,8 @@ import { MigrationClient } from '../../migration-client';
 import pino from 'pino';
 import { getChildLogger } from '../../../helpers';
 import { Inject, Service } from 'typedi';
-import { COLLECTION, LOGGER } from '../../../constants';
+import { LOGGER } from '../../../constants';
+import { COLLECTION } from '../constants';
 import deepmerge from 'deepmerge';
 import { createOne, deleteOne, readMany, updateOne } from './requests';
 import { SchemaClient, Type } from '../global';
