+++ packages/cli/src/lib/services/seed/collections/data-differ.ts
@@ -4,7 +4,8 @@ import { SeedDataClient } from './data-client';
 import { SeedDataMapper } from './data-mapper';
 import pino from 'pino';
 import { Inject, Service } from 'typedi';
-import { COLLECTION, LOGGER, META } from '../../../constants';
+import { LOGGER } from '../../../constants';
+import { COLLECTION, META } from '../constants';
 import { getChildLogger } from '../../../helpers';
 import { diff } from 'deep-object-diff';
 import { SeedMeta } from '../interfaces';
