+++ packages/cli/src/lib/services/seed/constants.ts
@@ -1,0 +1,5 @@
+export const COLLECTION = 'COLLECTION';
+export const META = 'META';
+
+export const DIRECTUS_COLLECTIONS_PREFIX = 'directus_';
+export const CUSTOM_COLLECTIONS_MAPPING_PREFIX = 'items:';
