+++ packages/cli/src/lib/services/seed/global/schema-client.ts
@@ -1,5 +1,6 @@
 import { Inject, Service } from 'typedi';
-import { LOGGER, DIRECTUS_COLLECTIONS_PREFIX } from '../../../constants';
+import { LOGGER } from '../../../constants';
+import { DIRECTUS_COLLECTIONS_PREFIX } from '../constants';
 import { getChildLogger } from '../../../helpers';
 import pino from 'pino';
 import { Cacheable } from 'typescript-cacheable';
